add a static bundle sidecar container build
This only adds the build process so we can start generating these images
regualarly, ahead of creating a trust-manager release which actually
uses the images.

Also adds buildx provisioning and uses named a named buildx builder
for trust-manager specifically.

The build is intended to be regularly scheduled in cloudbuild using
the provided cloudbuild.yaml

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>