subtle: change how flags are passed to smoke tests
This never worked previously because it was set up wrong.

This is documented on the ginkgo site but it's not obvious:

https://onsi.github.io/ginkgo/\#supporting-custom-configuration-custom-command-line-flags

This also moves the init to the "smoke_test.go" file rather than having
it in "smoke.go". This was required for the flags to be respected, and
in any case "smoke.go" just seems confusing when this directory /
package is entirely about testing.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>