add new alias gen ensuring unique JKS aliases
Adds tests of the bug we're fixing[1] and an integration
test for JKS functionality too.

Also exports default password and changes function signatures
to accept the password as a parameter.

This should make it easier to add support for custom passwords
in the future, if we decide to do so.

[1] https://github.com/cert-manager/trust-manager/pull/122#issuecomment-1529783592

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>