vendor dependencies correctly
Make won't update files when a variable changes unless that variable is
somehow included as a dependency for a target or in the path of the name
of the target.

This commit adds versions to the paths of all tools we depend on, so
that when one user bumps version numbers, everyone else won't need to
run `make clean`.

Signed-off-by: Ashley Davis <ashley.davis@venafi.com>