fix passwordless pkcs12 files for Java
A user discovered that Java's keytool was unable to read PKCS#12 files
generated by trust-manager which had no password set. We've not dug into
why, but presumably keytool's implementation treats an empty password
the same as a missing password somehow.

This wasn't caught in tests because our PKCS#12 library can parse the
resulting files just fine.

By changing to use the "passwordless" encoder when no password is
present, we can create files which both our tests and keytool can
read. It's possible that some other tools can't handle the passwordless
encoder and could break, but Java is (anecdotally) what we'd expect to
be the largest consumer of PKCS#12 trust stores, so we'll optimise for
that.

Signed-off-by: Ashley Davis <ashley.davis@venafi.com>